Bring it as close as possible to shell version
I tried to bring the ansible role as close as possible to the working
shell script deploy-seafile-pro-with-podman.sh which is now included
as a reference in this role.

It's possible to deploy a running version with the shell script but
not with the ansible role. Using the role login is possible but after
login you end up with a page unavailable error. Reason unknown.

Guess I need some help here.

Signed-off-by: Joerg Kastning <jkastning@my-it-brain.de>